refactor: Rename `dynamicSlices` > `slices` as we use these for fixed containers too
